make AtomGroup.coordinates() return an array of type numpy.float32 by default - distance array needs float32 and when interfacing with I/O this is also   how coordinates are stored so it makes sense to enforce it at this   crucial entry point - LeafletFinder would raise an error when building the adjacency matrix   because the array returned by coordinates() was float64
git-svn-id: https://mdanalysis.googlecode.com/svn/trunk@420 dcee7792-c13e-0410-88f5-272e5f544b36

+++ MDAnalysis/core/AtomGroup.py
@@ -301,10 +301,10 @@ class AtomGroup(object):
                       category=DeprecationWarning)
         return self.principalAxes().T
 
-    def coordinates(self, ts=None, copy=False):
+    def coordinates(self, ts=None, copy=False, dtype=numpy.float32):
         if ts == None:
             ts = self.universe.coord
-        return numpy.array(ts[self.indices()], copy=copy)
+        return numpy.array(ts[self.indices()], copy=copy, dtype=dtype)
 
     def selectAtoms(self, sel, *othersel):
         import Selection
